requirements.txt
certifi==2017.7.27.1
chardet==3.0.4
facebook-sdk==2.0.0
idna==2.5
requests==2.18.3
selenium==3.4.3
urllib3==1.22
